chore: separate operations to different contexts
